Rename AttachmentSourcePickerView -> AttachmentTypePickerView
